Adds BoardGameGeek rating to hot games
Adds a new 'bgg_rating' field to the BGG hot games data model and API schema.

This field represents the average rating of a game on BoardGameGeek and is populated during the data extraction process.

The addition of the bgg_rating allows for more comprehensive information about hot games to be displayed and analyzed.