Handles BGG login API inconsistencies
The BGG login API sometimes returns a 204 No Content
status code on success, instead of 200.

This change adapts the login process to handle both 200 and 204
status codes. It also addresses scenarios where the SessionID
cookie is not immediately available in the cookie jar and attempts
to extract it from the Set-Cookie headers. This covers inconsistencies
arising from different httpx versions.