Adds BGG purchase stats and API endpoints
Introduces API endpoints to retrieve purchase information and aggregated statistics for the user's BGG collection.

Adds endpoints for retrieving purchase history with pagination, a non-paginated endpoint for internal use, and an endpoint for aggregated purchase statistics (currency, status, type).

Updates scraper to ensure that it works with the DB of private data.